+++ components/collection/activity/ownerInsightsTabs/HolderTab.vue
@@ -74,7 +74,7 @@
 </template>
 
 <script setup lang="ts">
-import ProfileLink from '@/components/rmrk/Profile/ProfileLink.vue'
+import ProfileLink from '@/components/Profile/ProfileLink.vue'
 import { Owners } from '@/composables/collectionActivity/types'
 import { NeoIcon } from '@kodadot1/brick'
