+++ components/rmrk/Gallery/EmotionModal.vue
@@ -35,7 +35,7 @@
 </template>
 
 <script lang="ts" setup>
-import ProfileLink from '@/components/rmrk/Profile/ProfileLink.vue'
+import ProfileLink from '@/components/Profile/ProfileLink.vue'
 import { NeoButton } from '@kodadot1/brick'
 
 const props = defineProps<{
